chore(components): move types from the component files to *.types.ts
