+++ packages/ui-kit/src/components/Counter/index.ts
@@ -1,1 +1,2 @@
 export { Container as Counter } from './Container'
+export * from './Counter.types'
