+++ packages/ui-kit/src/components/Leaderboard/Leaderboard.tsx
@@ -10,20 +10,11 @@ import {
   Tooltip
 } from 'chart.js'
 import React from 'react'
-import {
-  customCanvasBackgroundColor,
-  DimensionInput,
-  FilterInput,
-  Propeller,
-  PROPEL_GRAPHQL_API_ENDPOINT,
-  Sort,
-  TimeRangeInput,
-  useLeaderboardQuery
-} from '../../helpers'
+import { customCanvasBackgroundColor, PROPEL_GRAPHQL_API_ENDPOINT, useLeaderboardQuery } from '../../helpers'
 import { ChartPlugins, ChartStyles, defaultChartHeight, defaultStyles } from '../../themes'
-import { ErrorFallback, ErrorFallbackProps } from '../ErrorFallback'
+import { ErrorFallback } from '../ErrorFallback'
 import { Loader } from '../Loader'
-import type { LeaderboardChartVariant, LeaderboardData } from './Leaderboard.types'
+import type { LeaderboardData } from './Leaderboard.types'
 import {
   getTableSettings,
   getValueWithPrefixAndSufix,
@@ -32,6 +23,7 @@ import {
   useSetupDefaultStyles
 } from './utils'
 import { ValueBar } from './ValueBar'
+import type { LeaderboardProps } from './Leaderboard.types'
 
 /**
  * It registers only the modules that will be used
@@ -41,55 +33,6 @@ import { ValueBar } from './ValueBar'
 ChartJS.register(BarController, BarElement, Tooltip, LinearScale, CategoryScale, Colors)
 
 let idCounter = 0
-
-export interface LeaderboardProps extends ErrorFallbackProps, React.ComponentProps<'canvas'> {
-  /** The variant the chart will respond to, can be either `bar` or `table` */
-  variant?: LeaderboardChartVariant
-
-  /** `styles` attribute can be either `BarStyles` or `TableStyles` */
-  styles?: ChartStyles
-
-  /** If passed along with `rows` the component will ignore the built-in graphql operations */
-  headers?: string[]
-
-  /** If passed along with `headers` the component will ignore the built-in graphql operations */
-  rows?: string[][]
-
-  /** When true, shows a skeleton loader */
-  loading?: boolean
-
-  query?: {
-    /** This should eventually be replaced to customer's app credentials */
-    accessToken?: string
-
-    /** Metric unique name */
-    metric?: string
-
-    /** Time range that the chart will respond to */
-    timeRange?: TimeRangeInput
-
-    /** Filters that the chart will respond to */
-    filters?: FilterInput[]
-
-    /** Propeller that the chart will respond to */
-    propeller?: Propeller
-
-    /** The number of rows to be returned. It can be a number between 1 and 1,000 */
-    rowLimit?: number
-
-    /** The sort order of the rows. It can be ascending (ASC) or descending (DESC) order. Defaults to descending (DESC) order when not provided. */
-    sort?: Sort
-
-    /** One or many Dimensions to group the Metric values by. Typically, Dimensions in a leaderboard are what you want to compare and rank. */
-    dimensions?: DimensionInput[]
-
-    /** Interval in milliseconds for refetching the data */
-    refetchInterval?: number
-
-    /** Whether to retry on errors. */
-    retry?: boolean
-  }
-}
 
 export function Leaderboard(props: LeaderboardProps) {
   const { variant = 'bar', styles, headers, rows, query, error, loading: isLoadingStatic = false, ...rest } = props
