+++ packages/ui-kit/src/components/TimeSeries/index.ts
@@ -1,2 +1,2 @@
 export { Container as TimeSeries } from './Container'
-export type { TimeSeriesChartVariant, TimeSeriesData } from './TimeSeries.types'
+export * from './TimeSeries.types'
